+++ src/blockchain/contracts/MockStableCoin.sol
@@ -7,14 +7,9 @@ import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
 import "@openzeppelin/contracts/access/Ownable.sol";
 
 contract MockStableCoin is Ownable, ERC20 {
-   
-    constructor() ERC20("MockStableCoin", "STABLE") {
-        
-    }
+    constructor() ERC20("MockStableCoin", "STABLE") {}
 
-    function Mint(address receiver, uint amount) public onlyOwner  {
+    function Mint(address receiver, uint256 amount) public onlyOwner {
         _mint(receiver, amount);
-
     }
-
 }
